implement intows for hyper request, cleanup and refactoring
