+++ src/server/upgrade/mod.rs
@@ -1,10 +1,36 @@
 //! Allows you to take an existing request or stream of data and convert it into a
 //! WebSocket client.
+extern crate hyper as real_hyper;
+
+use std::error::Error;
 use std::net::TcpStream;
-use std::io::Read;
+use std::io;
+use std::fmt::{
+	  Formatter,
+	  Display,
+	  self,
+};
 use stream::{
-	Stream,
-	AsTcpStream,
+	  Stream,
+	  AsTcpStream,
+};
+use header::{
+	  WebSocketKey,
+	  WebSocketVersion,
+};
+
+pub use self::real_hyper::http::h1::Incoming;
+pub use self::real_hyper::method::Method;
+pub use self::real_hyper::version::HttpVersion;
+pub use self::real_hyper::uri::RequestUri;
+pub use self::real_hyper::buffer::BufReader;
+pub use self::real_hyper::http::h1::parse_request;
+pub use self::real_hyper::header::{
+	  Headers,
+	  Upgrade,
+	  ProtocolName,
+	  Connection,
+	  ConnectionOption,
 };
 
 pub mod hyper;
@@ -16,34 +42,34 @@ pub mod hyper;
 /// Users should then call `accept` or `deny` to complete the handshake
 /// and start a session.
 pub struct WsUpgrade<S>
-where S: Stream,
+    where S: Stream,
 {
-	stream: S,
-	request: hyper::Request,
+	  stream: S,
+	  request: Request,
 }
 
 impl<S> WsUpgrade<S>
-where S: Stream,
+    where S: Stream,
 {
-	pub fn accept(self) {
-		unimplemented!();
-	}
+	  pub fn accept(self) {
+		    unimplemented!();
+	  }
 
-	pub fn reject(self) -> S {
-		unimplemented!();
-	}
+	  pub fn reject(self) -> S {
+		    unimplemented!();
+	  }
 
-	pub fn into_stream(self) -> S {
-		unimplemented!();
-	}
+	  pub fn into_stream(self) -> S {
+		    unimplemented!();
+	  }
 }
 
 impl<S> WsUpgrade<S>
-where S: Stream + AsTcpStream,
+    where S: Stream + AsTcpStream,
 {
-	pub fn tcp_stream(&self) -> &TcpStream {
-		self.stream.as_tcp()
-	}
+	  pub fn tcp_stream(&self) -> &TcpStream {
+		    self.stream.as_tcp()
+	  }
 }
 
 /// Trait to take a stream or similar and attempt to recover the start of a
@@ -56,11 +82,176 @@ where S: Stream + AsTcpStream,
 ///
 /// Note: the stream is owned because the websocket client expects to own its stream.
 pub trait IntoWs {
-	type Stream: Stream;
-	type Error;
-	/// Attempt to parse the start of a Websocket handshake, later with the  returned
-	/// `WsUpgrade` struct, call `accept to start a websocket client, and `reject` to
-	/// send a handshake rejection response.
-	fn into_ws(mut self) -> Result<WsUpgrade<Self::Stream>, Self::Error>;
+	  type Stream: Stream;
+	  type Error;
+	  /// Attempt to parse the start of a Websocket handshake, later with the  returned
+	  /// `WsUpgrade` struct, call `accept to start a websocket client, and `reject` to
+	  /// send a handshake rejection response.
+	  fn into_ws(self) -> Result<WsUpgrade<Self::Stream>, Self::Error>;
 }
 
+
+pub type Request = Incoming<(Method, RequestUri)>;
+pub struct RequestStreamPair<S: Stream>(pub S, pub Request);
+
+impl<S> IntoWs for S
+    where S: Stream,
+{
+	  type Stream = S;
+	  type Error = (Self, Option<Request>, HyperIntoWsError);
+
+	  fn into_ws(mut self) -> Result<WsUpgrade<Self::Stream>, Self::Error> {
+		    let request = {
+			      let mut reader = BufReader::new(self.reader());
+			      parse_request(&mut reader)
+		    };
+
+		    let request = match request {
+			      Ok(r) => r,
+			      Err(e) => return Err((self, None, e.into())),
+		    };
+
+		    match validate(&request.subject.0, &request.version, &request.headers) {
+			      Ok(_) => Ok(WsUpgrade {
+				        stream: self,
+				        request: request,
+			      }),
+			      Err(e) => Err((self, Some(request), e)),
+		    }
+	  }
+}
+
+impl<S> IntoWs for RequestStreamPair<S>
+    where S: Stream,
+{
+	  type Stream = S;
+	  type Error = (S, Request, HyperIntoWsError);
+
+	  fn into_ws(self) -> Result<WsUpgrade<Self::Stream>, Self::Error> {
+		    match validate(&self.1.subject.0, &self.1.version, &self.1.headers) {
+			      Ok(_) => Ok(WsUpgrade {
+				        stream: self.0,
+				        request: self.1,
+			      }),
+			      Err(e) => Err((self.0, self.1, e)),
+		    }
+	  }
+}
+
+#[derive(Debug)]
+pub enum HyperIntoWsError {
+	  MethodNotGet,
+	  UnsupportedHttpVersion,
+	  UnsupportedWebsocketVersion,
+	  NoSecWsKeyHeader,
+	  NoWsUpgradeHeader,
+	  NoUpgradeHeader,
+	  NoWsConnectionHeader,
+	  NoConnectionHeader,
+	  UnknownNetworkStream,
+	  /// IO error from reading the underlying socket
+	  Io(io::Error),
+	  /// Error while parsing an incoming request
+	  Parsing(self::real_hyper::error::Error),
+}
+
+impl Display for HyperIntoWsError {
+	  fn fmt(&self, fmt: &mut Formatter) -> Result<(), fmt::Error> {
+		    fmt.write_str(self.description())
+	  }
+}
+
+impl Error for HyperIntoWsError {
+	  fn description(&self) -> &str {
+		    use self::HyperIntoWsError::*;
+		    match self {
+			      &MethodNotGet => "Request method must be GET",
+			      &UnsupportedHttpVersion => "Unsupported request HTTP version",
+			      &UnsupportedWebsocketVersion => "Unsupported WebSocket version",
+			      &NoSecWsKeyHeader => "Missing Sec-WebSocket-Key header",
+			      &NoWsUpgradeHeader => "Invalid Upgrade WebSocket header",
+			      &NoUpgradeHeader => "Missing Upgrade WebSocket header",
+			      &NoWsConnectionHeader => "Invalid Connection WebSocket header",
+			      &NoConnectionHeader => "Missing Connection WebSocket header",
+			      &UnknownNetworkStream => "Cannot downcast to known impl of NetworkStream",
+			      &Io(ref e) => e.description(),
+			      &Parsing(ref e) => e.description(),
+		    }
+	  }
+
+	  fn cause(&self) -> Option<&Error> {
+		    match *self {
+			      HyperIntoWsError::Io(ref e) => Some(e),
+			      HyperIntoWsError::Parsing(ref e) => Some(e),
+			      _ => None,
+		    }
+	  }
+}
+
+impl From<io::Error> for HyperIntoWsError {
+	  fn from(err: io::Error) -> Self {
+		    HyperIntoWsError::Io(err)
+	  }
+}
+
+impl From<real_hyper::error::Error> for HyperIntoWsError {
+	  fn from(err: real_hyper::error::Error) -> Self {
+		    HyperIntoWsError::Parsing(err)
+	  }
+}
+
+pub fn validate(
+	  method: &Method,
+	  version: &HttpVersion,
+	  headers: &Headers
+) -> Result<(), HyperIntoWsError>
+{
+	  if *method != Method::Get {
+		    return Err(HyperIntoWsError::MethodNotGet);
+	  }
+
+	  if *version == HttpVersion::Http09 || *version == HttpVersion::Http10 {
+		    return Err(HyperIntoWsError::UnsupportedHttpVersion);
+	  }
+
+	  if let Some(version) = headers.get::<WebSocketVersion>() {
+		    if version != &WebSocketVersion::WebSocket13 {
+			      return Err(HyperIntoWsError::UnsupportedWebsocketVersion);
+		    }
+	  }
+
+	  if headers.get::<WebSocketKey>().is_none() {
+		    return Err(HyperIntoWsError::NoSecWsKeyHeader);
+	  }
+
+	  match headers.get() {
+		    Some(&Upgrade(ref upgrade)) => {
+			      if upgrade.iter().all(|u| u.name != ProtocolName::WebSocket) {
+				        return Err(HyperIntoWsError::NoWsUpgradeHeader)
+			      }
+		    },
+		    None => return Err(HyperIntoWsError::NoUpgradeHeader),
+	  };
+
+	  fn check_connection_header(headers: &Vec<ConnectionOption>) -> bool {
+		    for header in headers {
+			      if let &ConnectionOption::ConnectionHeader(ref h) = header {
+				        if h as &str == "upgrade" {
+					          return true;
+				        }
+			      }
+		    }
+		    false
+	  }
+
+	  match headers.get() {
+		    Some(&Connection(ref connection)) => {
+			      if !check_connection_header(connection) {
+				        return Err(HyperIntoWsError::NoWsConnectionHeader);
+			      }
+		    },
+		    None => return Err(HyperIntoWsError::NoConnectionHeader),
+	  };
+
+	  Ok(())
+}
